Avoid parallelization if not worthy (#6)
* Avoid parallelization if not worthy

In principle parallelizing operations is fine, but one has to consider
that there are overheads associated with communications, and sometimes
these can become too big.

In the particular case of having a single core for our parallelization
there clearly is no benefit on trying to parallelize anything, as we
simply add overheads to the system. Hence it makes sense to perform
serial operations in such cases.

This commit changes the logic of the two main always-parallelized
routines so that they take into account the number of cores at their
disposal before taking the decision of doing things in parallel or in
serial. At the moment the condition they check is very simple (do I have
more than 1 core at my disposal?) but in reality the balance will
probably lie somewhere else depending on the cost of the computation and
the available resources. This change alone takes down the runtime of
tests in my machine from ~22/23 minutes to ~3, a big gain. Examples also
about an order of magnitude faster.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

* Bring back declaration for i to avoid NOTE

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>